add search form and some small refactoring
