+++ cms_form/models/__init__.py
@@ -1,3 +1,5 @@
+from . import cms_form_mixin
 from . import cms_form
+from . import cms_search_form
 from . import test_models
 from . import website_mixin
